Refactor user and wardrobe interfaces and controller functions
This commit refactors the user and wardrobe interfaces and updates the corresponding controller functions to use camel case naming convention for the properties. The changes include renaming properties in the interfaces from snake case to camel case, such as `user_id` to `userId` and `is_admin` to `isAdmin`. The controller functions have also been updated to use the renamed properties. This improves code readability and consistency.